feat: 滚轮调整大小逻辑移动到 useModel
